+++ changelog.md
@@ -1,3 +1,8 @@
+### [v0.15.2.3] - 2026-06-17
+1. **Import Traveller World Systems:** Fixed import slot error resulting in system viewer mis-aligning orbits
+2. **System Viewer:** Now distinguishes between Size 0 belts and Size 0 worldlets
+3. **System Viewer:** Options added to hide moons, habitable zone, mainworld highlights
+
 ### [v0.15.2.2] - 2026-06-16
 1. **Import Traveller World Systems:** Fixed import error changing Orbit #